adição da variável de controle sobre qual procedimento executar

See merge request inovisao/compara_classificadores_torch!10

+++ run/rodaCruzadaBalanceado.sh
@@ -10,6 +10,9 @@ ndobras=5
 rodaPadrao=true
 rodaSiamesa=true
 
+# "treino" para apenas treinar, "teste" para apenas testar ou "completo" para executar ambos
+procedimento="completo"
+
 # Verifica se o usuário passou como parâmetro
 # o número de dobras (E.g.: ./rodaCruzada.sh -k 5)
 while getopts "k:" flag; 
@@ -34,8 +37,11 @@ for((i=1;i<=$ndobras;i+=1)); do folds+=("fold_${i}"); done
 mkdir -p ../results_dl/           
 rm -rf ../results_dl/*
 
-mkdir -p ../model_checkpoints/
-rm -rf ../model_checkpoints/*
+if [ "$procedimento" != "teste" ]
+then
+   mkdir -p ../model_checkpoints/
+   rm -rf ../model_checkpoints/*
+fi
 
 echo  'run,learning_rate,architecture,optimizer,precision,recall,fscore' > ../results_dl/results.csv
 
@@ -78,7 +84,7 @@ do
    mkdir -p ../results/history
    mkdir -p ../results/matrix
 
-   bash ./roda.sh $run $rodaPadrao $rodaSiamesa
+   bash ./roda.sh $run $rodaPadrao $rodaSiamesa $procedimento
   
    mkdir -p ${pastaDobrasResultados}/${Teste}
    mv ${pastaResultados}/* ${pastaDobrasResultados}/${Teste} 
@@ -86,8 +92,10 @@ do
 
 done
 
-cd ../src
-Rscript ./graphics.R
-cd ../run
+if [ "$procedimento" != "treino" ]
+then
+   cd ../src
+   Rscript ./graphics.R
+   cd ../run
+fi
  
-
